chapter6/strands-handson/backend/requirements.txt
strands-agents==1.1.0
strands-agents-tools==0.2.3
bedrock-agentcore==0.1.1
uv==0.8.4